requirements.txt
# Core data science libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Time series specific libraries
statsmodels>=0.14.0
pmdarima>=2.0.0
prophet>=1.1.0
tslearn>=0.6.0
darts>=0.24.0
sktime>=0.21.0

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Anomaly detection
pyod>=1.1.0

# Web interface
streamlit>=1.25.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
